Emit and consume map key for structs ...
As a consequence of calling `_emit_map_len()` maps are encoded in the msgpack
stream as `fixmap`.

This map cannot be consumed by other (non-Rust) msgpack decoders because the
keys are not present in this map; so it will end up decoding as one long stream
of binary data representing the map.

If structs are going to be encoded using the msgpack map type then they
should incl. (and be able to consume) keys in the msgpack stream.